Rename folder tf_baselines and remove setup.py .
