+++ og_marl/tf2_baselines/systems/iddpg_cql.py
@@ -12,27 +12,26 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
-"""Implementation of OMAR"""
+"""Implementation of IDDPG+CQL"""
 from typing import Any, Dict, Optional, Tuple
-import copy
 
+import copy
 import hydra
-import tree
 import numpy as np
 from omegaconf import DictConfig
 import tensorflow as tf
-import sonnet as snt
 from tensorflow import Tensor
-import tensorflow_probability as tfp
+import sonnet as snt
+import tree
 from chex import Numeric
 
 from og_marl.environments import get_environment, BaseEnvironment
 from og_marl.loggers import BaseLogger, WandbLogger
 from og_marl.vault_utils.download_vault import download_and_unzip_vault
 from og_marl.replay_buffers import Experience, FlashbaxReplayBuffer
-from og_marl.tf2.networks import StateAndActionCritic
-from og_marl.tf2.systems.base import BaseOfflineSystem
-from og_marl.tf2.utils import (
+from og_marl.tf2_baselines.networks import StateAndActionCritic
+from og_marl.tf2_baselines.systems.base import BaseOfflineSystem
+from og_marl.tf2_baselines.utils import (
     batch_concat_agent_id_to_obs,
     concat_agent_id_to_obs,
     expand_batch_and_agent_dim_of_time_major_sequence,
@@ -45,7 +44,13 @@ from og_marl.tf2.utils import (
 set_growing_gpu_memory()
 
 
-class OMARSystem(BaseOfflineSystem):
+class IDDPGCQLSystem(BaseOfflineSystem):
+    """Independent DDPG + CQL.
+
+    NOTE: the critic conditions on states and individual agent actions.
+
+    """
+
     def __init__(
         self,
         environment: BaseEnvironment,
@@ -60,19 +65,14 @@ class OMARSystem(BaseOfflineSystem):
         num_ood_actions: int = 10,  # CQL
         cql_weight: float = 3.0,  # CQL
         cql_sigma: float = 0.3,  # CQL
-        omar_iters: int = 3,  # OMAR
-        omar_num_samples: int = 10,  # OMAR
-        omar_num_elites: int = 10,  # OMAR
-        omar_sigma: float = 2.0,  # OMAR
-        omar_coe: float = 0.7,  # OMAR
     ):
         super().__init__(
             environment=environment,
             logger=logger,
         )
 
-        self.add_agent_id_to_obs = add_agent_id_to_obs
         self.discount = discount
+        self.add_agent_id_to_obs = add_agent_id_to_obs
 
         # Policy network
         self.policy_network = snt.DeepRNN(
@@ -109,18 +109,9 @@ class OMARSystem(BaseOfflineSystem):
             agent: self.policy_network.initial_state(1) for agent in self.environment.agents
         }
 
-        # CQL
         self.num_ood_actions = num_ood_actions
         self.cql_weight = cql_weight
         self.cql_sigma = cql_sigma
-
-        # OMAR
-        self.omar_iters = omar_iters
-        self.omar_num_samples = omar_num_samples
-        self.omar_num_elites = omar_num_elites
-        self.omar_coe = omar_coe
-
-        self.init_omar_mu, self.init_omar_sigma = 0.0, omar_sigma
 
     def reset(self) -> None:
         """Called at the start of a new episode."""
@@ -172,10 +163,12 @@ class OMARSystem(BaseOfflineSystem):
     @tf.function(jit_compile=True)  # NOTE: comment this out if using debugger
     def _tf_train_step(self, experience: Dict[str, Any]) -> Dict[str, Numeric]:
         # Unpack the batch
-        observations = experience["observations"]  # (B,T,N,O)
-        actions = experience["actions"]  # (B,T,N,A)
-        env_states = experience["infos"]["state"]  # (B,T,S)
-        rewards = experience["rewards"]  # (B,T,N)
+        observations = tf.cast(experience["observations"], "float32")  # (B,T,N,O)
+        actions = tf.cast(
+            tf.clip_by_value(experience["actions"], -1.0, 1.0), "float32"
+        )  # (B,T,N,A)
+        env_states = tf.cast(experience["infos"]["state"], "float32")  # (B,T,S)
+        rewards = tf.cast(experience["rewards"], "float32")  # (B,T,N)
         truncations = tf.cast(experience["truncations"], "float32")  # (B,T,N)
         terminals = tf.cast(experience["terminals"], "float32")  # (B,T,N)
 
@@ -183,7 +176,7 @@ class OMARSystem(BaseOfflineSystem):
         resets = tf.maximum(terminals, truncations)  # equivalent to logical 'or'
 
         # Get dims
-        B, T, N, A = actions.shape[:4]
+        B, T, N = actions.shape[:3]
 
         # Maybe add agent ids to observation
         if self.add_agent_id_to_obs:
@@ -191,11 +184,11 @@ class OMARSystem(BaseOfflineSystem):
 
         # Make time-major
         observations = switch_two_leading_dims(observations)
-        resets = switch_two_leading_dims(resets)
         replay_actions = switch_two_leading_dims(actions)
         rewards = switch_two_leading_dims(rewards)
         terminals = switch_two_leading_dims(terminals)
         env_states = switch_two_leading_dims(env_states)
+        resets = switch_two_leading_dims(resets)
 
         # Unroll target policy
         target_actions = unroll_rnn(
@@ -233,16 +226,28 @@ class OMARSystem(BaseOfflineSystem):
             critic_loss_1 = tf.reduce_mean(0.5 * (targets - qs_1[:-1]) ** 2)
             critic_loss_2 = tf.reduce_mean(0.5 * (targets - qs_2[:-1]) ** 2)
 
-            ###########
-            ### CQL ###
-            ###########
-
+            ###################
+            ### Policy Loss ###
+            ###################
             online_actions = unroll_rnn(
                 self.policy_network,
                 merge_batch_and_agent_dim_of_time_major_sequence(observations),
                 merge_batch_and_agent_dim_of_time_major_sequence(resets),
             )
             online_actions = expand_batch_and_agent_dim_of_time_major_sequence(online_actions, B, N)
+
+            # Unroll online policy
+            policy_qs_1 = self.critic_network_1(env_states, online_actions)
+            policy_qs_2 = self.critic_network_2(env_states, online_actions)
+            policy_qs = tf.minimum(policy_qs_1, policy_qs_2)
+
+            policy_loss = -tf.reduce_mean(policy_qs) + 1e-3 * tf.reduce_mean(
+                tf.square(online_actions)
+            )
+
+            ###########
+            ### CQL ###
+            ###########
 
             # Repeat all tensors num_ood_actions times andadd  next to batch dim
             repeat_observations = tf.stack(
@@ -345,102 +350,20 @@ class OMARSystem(BaseOfflineSystem):
             all_ood_qs_1 = tf.concat((ood_qs_1, current_ood_qs_1, next_current_ood_qs_1), axis=2)
             all_ood_qs_2 = tf.concat((ood_qs_2, current_ood_qs_2, next_current_ood_qs_2), axis=2)
 
-            cql_loss_1 = tf.reduce_mean(
-                tf.reduce_logsumexp(all_ood_qs_1, axis=2, keepdims=False)
-            ) - tf.reduce_mean(qs_1[:-1])
-            cql_loss_2 = tf.reduce_mean(
-                tf.reduce_logsumexp(all_ood_qs_2, axis=2, keepdims=False)
-            ) - tf.reduce_mean(qs_2[:-1])
+            cql_loss_1 = tf.reduce_mean(tf.reduce_logsumexp(all_ood_qs_1, axis=2)) - tf.reduce_mean(
+                qs_1[:-1]
+            )
+            cql_loss_2 = tf.reduce_mean(tf.reduce_logsumexp(all_ood_qs_2, axis=2)) - tf.reduce_mean(
+                qs_2[:-1]
+            )
 
             critic_loss_1 += self.cql_weight * cql_loss_1
             critic_loss_2 += self.cql_weight * cql_loss_2
 
             ### END CQL ###
 
+            # Combine critic loss
             critic_loss = (critic_loss_1 + critic_loss_2) / 2
-
-            ####################
-            # OMAR Policy Loss #
-            ####################
-
-            curr_pol_out = online_actions
-
-            pred_qvals = tf.minimum(
-                self.critic_network_1(env_states, curr_pol_out),
-                self.critic_network_2(env_states, curr_pol_out),
-            )
-
-            omar_mu = tf.zeros_like(curr_pol_out) + self.init_omar_mu
-            omar_sigma = tf.zeros_like(curr_pol_out) + self.init_omar_sigma
-
-            # Repeat all tensors num_ood_actions times andadd  next to batch dim
-            observations = tf.stack(
-                [observations] * self.omar_num_samples, axis=2
-            )  # next to batch dim
-            env_states = tf.stack([env_states] * self.omar_num_samples, axis=2)  # next to batch dim
-
-            # Flatten into batch dim
-            observations = tf.reshape(observations, (T, -1, *observations.shape[3:]))
-            env_states = tf.reshape(env_states, (T, -1, *env_states.shape[3:]))
-
-            last_top_k_qvals, last_elite_acs = None, None
-            for iter_idx in range(self.omar_iters):
-                dist = tfp.distributions.Normal(omar_mu, omar_sigma)
-                cem_sampled_acs = dist.sample((self.omar_num_samples,))
-                cem_sampled_acs = tf.transpose(cem_sampled_acs, (1, 2, 0, 3, 4))
-                cem_sampled_acs = tf.clip_by_value(cem_sampled_acs, -1.0, 1.0)
-
-                formatted_cem_sampled_acs = tf.reshape(
-                    cem_sampled_acs, (T, -1, *cem_sampled_acs.shape[3:])
-                )
-
-                all_pred_qvals = tf.minimum(
-                    self.critic_network_1(env_states, formatted_cem_sampled_acs),
-                    self.critic_network_2(env_states, formatted_cem_sampled_acs),
-                )
-                all_pred_qvals = tf.reshape(all_pred_qvals, (T, B, self.omar_num_samples, N))
-                all_pred_qvals = tf.transpose(all_pred_qvals, (0, 1, 3, 2))
-                cem_sampled_acs = tf.transpose(cem_sampled_acs, (0, 1, 3, 4, 2))
-
-                if iter_idx > 0:
-                    all_pred_qvals = tf.concat((all_pred_qvals, last_top_k_qvals), axis=-1)  # type: ignore
-                    cem_sampled_acs = tf.concat((cem_sampled_acs, last_elite_acs), axis=-1)  # type: ignore
-
-                top_k_qvals, top_k_inds = tf.math.top_k(all_pred_qvals, self.omar_num_elites)
-                elite_ac_inds = tf.stack([top_k_inds] * A, axis=-2)
-                elite_acs = tf.gather(cem_sampled_acs, elite_ac_inds, batch_dims=-1)
-
-                last_top_k_qvals, last_elite_acs = top_k_qvals, elite_acs
-
-                updated_mu = tf.reduce_mean(elite_acs, axis=-1)
-                updated_sigma = tf.math.reduce_std(elite_acs, axis=-1)
-
-                omar_mu = updated_mu
-                omar_sigma = updated_sigma
-
-            top_qvals, top_inds = tf.math.top_k(all_pred_qvals, self.omar_num_elites)
-            top_ac_inds = tf.stack([top_k_inds] * A, axis=-2)
-            top_acs = tf.gather(cem_sampled_acs, top_ac_inds, batch_dims=-1)
-
-            cem_qvals = top_qvals
-            pol_qvals = pred_qvals
-            cem_acs = top_acs
-            pol_acs = tf.expand_dims(curr_pol_out, axis=-1)
-
-            candidate_qvals = tf.concat([pol_qvals, cem_qvals], -1)
-            candidate_acs = tf.concat([pol_acs, cem_acs], -1)
-
-            max_inds = tf.argmax(candidate_qvals, axis=-1)
-            max_ac_inds = tf.expand_dims(tf.stack([max_inds] * A, axis=-1), axis=-1)
-
-            max_acs = tf.gather(candidate_acs, max_ac_inds, batch_dims=-1)
-            max_acs = tf.stop_gradient(tf.reshape(max_acs, max_acs.shape[:-1]))
-
-            policy_loss = (
-                self.omar_coe * tf.reduce_mean((curr_pol_out - max_acs) ** 2)
-                - (1 - self.omar_coe) * tf.reduce_mean(pred_qvals)
-                + 1e-3 * tf.reduce_mean(curr_pol_out**2)
-            )
 
         # Train critics
         variables = (
@@ -479,14 +402,15 @@ class OMARSystem(BaseOfflineSystem):
         logs = {
             "mean_dataset_q_values": tf.reduce_mean((qs_1 + qs_2) / 2),
             "critic_loss": critic_loss,
-            "cql_loss": tf.reduce_mean(cql_loss_1 + cql_loss_2) / 2.0,
+            "cql_loss": (cql_loss_1 + cql_loss_2) / 2.0,
             "policy_loss": policy_loss,
+            "mean_chosen_q_values": tf.reduce_mean((policy_qs_1 + policy_qs_2) / 2),
         }
 
         return logs
 
 
-@hydra.main(version_base=None, config_path="configs", config_name="omar")
+@hydra.main(version_base=None, config_path="configs", config_name="iddpg_cql")
 def run_experiment(cfg: DictConfig) -> None:
     print(cfg)
 
@@ -512,7 +436,7 @@ def run_experiment(cfg: DictConfig) -> None:
     }
     logger = WandbLogger(project=cfg["wandb_project"], config=wandb_config)
 
-    system = OMARSystem(env, logger, **cfg["system"])
+    system = IDDPGCQLSystem(env, logger, **cfg["system"])
 
     tf.random.set_seed(cfg["seed"])
