+++ og_marl/tf2_baselines/systems/maddpg_cql.py
@@ -12,26 +12,26 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
-"""Implementation of IDDPG+CQL"""
+"""Implementation of MADDPG+CQL"""
 from typing import Any, Dict, Optional, Tuple
-
 import copy
+
 import hydra
 import numpy as np
+import sonnet as snt
 from omegaconf import DictConfig
 import tensorflow as tf
 from tensorflow import Tensor
-import sonnet as snt
 import tree
 from chex import Numeric
 
 from og_marl.environments import get_environment, BaseEnvironment
+from og_marl.tf2_baselines.networks import StateAndJointActionCritic
+from og_marl.tf2_baselines.systems.base import BaseOfflineSystem
 from og_marl.loggers import BaseLogger, WandbLogger
 from og_marl.vault_utils.download_vault import download_and_unzip_vault
 from og_marl.replay_buffers import Experience, FlashbaxReplayBuffer
-from og_marl.tf2.networks import StateAndActionCritic
-from og_marl.tf2.systems.base import BaseOfflineSystem
-from og_marl.tf2.utils import (
+from og_marl.tf2_baselines.utils import (
     batch_concat_agent_id_to_obs,
     concat_agent_id_to_obs,
     expand_batch_and_agent_dim_of_time_major_sequence,
@@ -44,12 +44,8 @@ from og_marl.tf2.utils import (
 set_growing_gpu_memory()
 
 
-class IDDPGBCSystem(BaseOfflineSystem):
-    """Independent DDPG + BC.
-
-    NOTE: the critic conditions on states and individual agent actions.
-
-    """
+class MADDPGCQLSystem(BaseOfflineSystem):
+    """Multi-Agent Deep Deterministic Policy Gradients with CQL."""
 
     def __init__(
         self,
@@ -59,18 +55,17 @@ class IDDPGBCSystem(BaseOfflineSystem):
         recurrent_layer_dim: int = 64,
         discount: float = 0.99,
         target_update_rate: float = 0.005,
-        critic_learning_rate: float = 3e-4,
+        critic_learning_rate: float = 1e-3,
         policy_learning_rate: float = 3e-4,
         add_agent_id_to_obs: bool = True,
-        bc_alpha: float = 2.5,  # BC
+        num_ood_actions: int = 10,  # CQL
+        cql_weight: float = 3.0,  # CQL
+        cql_sigma: float = 0.2,  # CQL
     ):
-        super().__init__(
-            environment=environment,
-            logger=logger,
-        )
+        super().__init__(environment=environment, logger=logger)
 
-        self.discount = discount
         self.add_agent_id_to_obs = add_agent_id_to_obs
+        self.discount = discount
 
         # Policy network
         self.policy_network = snt.DeepRNN(
@@ -88,8 +83,8 @@ class IDDPGBCSystem(BaseOfflineSystem):
         self.target_policy_network = copy.deepcopy(self.policy_network)
 
         # Critic network
-        self.critic_network_1 = StateAndActionCritic(
-            len(self.environment.agents), self.environment.num_actions, add_agent_id_to_obs
+        self.critic_network_1 = StateAndJointActionCritic(
+            len(self.environment.agents), self.environment.num_actions
         )  # shared network for all agents
         self.critic_network_2 = copy.deepcopy(self.critic_network_1)
 
@@ -107,7 +102,10 @@ class IDDPGBCSystem(BaseOfflineSystem):
             agent: self.policy_network.initial_state(1) for agent in self.environment.agents
         }
 
-        self.bc_alpha = bc_alpha
+        # CQL
+        self.num_ood_actions = num_ood_actions
+        self.cql_weight = cql_weight
+        self.cql_sigma = cql_sigma
 
     def reset(self) -> None:
         """Called at the start of a new episode."""
@@ -159,12 +157,12 @@ class IDDPGBCSystem(BaseOfflineSystem):
     @tf.function(jit_compile=True)  # NOTE: comment this out if using debugger
     def _tf_train_step(self, experience: Dict[str, Any]) -> Dict[str, Numeric]:
         # Unpack the batch
-        observations = tf.cast(experience["observations"], "float32")  # (B,T,N,O)
-        replay_actions = tf.cast(
-            tf.clip_by_value(experience["actions"], -1.0, 1.0), "float32"
-        )  # (B,T,N,A)
-        env_states = tf.cast(experience["infos"]["state"], "float32")  # (B,T,S)
-        rewards = tf.cast(experience["rewards"], "float32")  # (B,T,N)
+        observations = experience["observations"]  # (B,T,N,O)
+        actions = tf.clip_by_value(
+            experience["actions"], -1.0, 1.0
+        )  # (B,T,N,A) clip for omiga datasets
+        env_states = experience["infos"]["state"]  # (B,T,S)
+        rewards = experience["rewards"]  # (B,T,N)
         truncations = tf.cast(experience["truncations"], "float32")  # (B,T,N)
         terminals = tf.cast(experience["terminals"], "float32")  # (B,T,N)
 
@@ -172,7 +170,7 @@ class IDDPGBCSystem(BaseOfflineSystem):
         resets = tf.maximum(terminals, truncations)  # equivalent to logical 'or'
 
         # Get dims
-        B, T, N = replay_actions.shape[:3]
+        B, T, N = actions.shape[:3]
 
         # Maybe add agent ids to observation
         if self.add_agent_id_to_obs:
@@ -180,7 +178,7 @@ class IDDPGBCSystem(BaseOfflineSystem):
 
         # Make time-major
         observations = switch_two_leading_dims(observations)
-        replay_actions = switch_two_leading_dims(replay_actions)
+        replay_actions = switch_two_leading_dims(actions)
         rewards = switch_two_leading_dims(rewards)
         terminals = switch_two_leading_dims(terminals)
         env_states = switch_two_leading_dims(env_states)
@@ -195,8 +193,8 @@ class IDDPGBCSystem(BaseOfflineSystem):
         target_actions = expand_batch_and_agent_dim_of_time_major_sequence(target_actions, B, N)
 
         # Target critics
-        target_qs_1 = self.target_critic_network_1(env_states, target_actions)
-        target_qs_2 = self.target_critic_network_2(env_states, target_actions)
+        target_qs_1 = self.target_critic_network_1(env_states, target_actions, target_actions)
+        target_qs_2 = self.target_critic_network_2(env_states, target_actions, target_actions)
 
         # Take minimum between two target critics
         target_qs = tf.minimum(target_qs_1, target_qs_2)
@@ -210,24 +208,22 @@ class IDDPGBCSystem(BaseOfflineSystem):
         with tf.GradientTape(persistent=True) as tape:
             # Online critics
             qs_1 = tf.squeeze(
-                self.critic_network_1(env_states, replay_actions),
+                self.critic_network_1(env_states, replay_actions, replay_actions),
                 axis=-1,
             )
             qs_2 = tf.squeeze(
-                self.critic_network_2(env_states, replay_actions),
+                self.critic_network_2(env_states, replay_actions, replay_actions),
                 axis=-1,
             )
 
             # Squared TD-error
-            critic_loss_1 = tf.reduce_mean(0.5 * (targets - qs_1[:-1]) ** 2)
-            critic_loss_2 = tf.reduce_mean(0.5 * (targets - qs_2[:-1]) ** 2)
+            td_error_1 = tf.reduce_mean(0.5 * (targets - qs_1[:-1]) ** 2)
+            td_error_2 = tf.reduce_mean(0.5 * (targets - qs_2[:-1]) ** 2)
 
-            # Combine critic loss
-            critic_loss = (critic_loss_1 + critic_loss_2) / 2
+            ###########
+            ### CQL ###
+            ###########
 
-            ###################
-            ### Policy Loss ###
-            ###################
             online_actions = unroll_rnn(
                 self.policy_network,
                 merge_batch_and_agent_dim_of_time_major_sequence(observations),
@@ -235,16 +231,134 @@ class IDDPGBCSystem(BaseOfflineSystem):
             )
             online_actions = expand_batch_and_agent_dim_of_time_major_sequence(online_actions, B, N)
 
-            # Unroll online policy
-            policy_qs_1 = self.critic_network_1(env_states, online_actions)
-            policy_qs_2 = self.critic_network_2(env_states, online_actions)
+            # Repeat all tensors num_ood_actions times andadd  next to batch dim
+            repeat_observations = tf.stack(
+                [observations] * self.num_ood_actions, axis=2
+            )  # next to batch dim
+            repeat_env_states = tf.stack(
+                [env_states] * self.num_ood_actions, axis=2
+            )  # next to batch dim
+            repeat_online_actions = tf.stack(
+                [online_actions] * self.num_ood_actions, axis=2
+            )  # next to batch dim
+
+            # Flatten into batch dim
+            repeat_observations = tf.reshape(
+                repeat_observations, (T, -1, *repeat_observations.shape[3:])
+            )
+            repeat_env_states = tf.reshape(repeat_env_states, (T, -1, *repeat_env_states.shape[3:]))
+            repeat_online_actions = tf.reshape(
+                repeat_online_actions, (T, -1, *repeat_online_actions.shape[3:])
+            )
+
+            # CQL Loss
+            random_ood_actions = tf.random.uniform(
+                shape=repeat_online_actions.shape,
+                minval=-1.0,
+                maxval=1.0,
+                dtype=repeat_online_actions.dtype,
+            )
+            random_ood_action_log_pi = tf.math.log(0.5 ** (random_ood_actions.shape[-1]))
+
+            ood_qs_1 = (
+                self.critic_network_1(repeat_env_states, random_ood_actions, random_ood_actions)[
+                    :-1
+                ]
+                - random_ood_action_log_pi
+            )
+            ood_qs_2 = (
+                self.critic_network_2(repeat_env_states, random_ood_actions, random_ood_actions)[
+                    :-1
+                ]
+                - random_ood_action_log_pi
+            )
+
+            # # Actions near true actions
+            mu = 0.0
+            std = self.cql_sigma
+            action_noise = tf.random.normal(
+                repeat_online_actions.shape,
+                mean=mu,
+                stddev=std,
+                dtype=repeat_online_actions.dtype,
+            )
+            current_ood_actions = tf.clip_by_value(repeat_online_actions + action_noise, -1.0, 1.0)
+
+            ood_actions_prob = (1 / (self.cql_sigma * tf.math.sqrt(2 * np.pi))) * tf.exp(
+                -((action_noise - mu) ** 2) / (2 * self.cql_sigma**2)
+            )
+            ood_actions_log_prob = tf.math.log(
+                tf.reduce_prod(ood_actions_prob, axis=-1, keepdims=True)
+            )
+
+            current_ood_qs_1 = (
+                self.critic_network_1(
+                    repeat_env_states[:-1],
+                    current_ood_actions[:-1],
+                    current_ood_actions[:-1],
+                )
+                - ood_actions_log_prob[:-1]
+            )
+            current_ood_qs_2 = (
+                self.critic_network_2(
+                    repeat_env_states[:-1],
+                    current_ood_actions[:-1],
+                    current_ood_actions[:-1],
+                )
+                - ood_actions_log_prob[:-1]
+            )
+
+            next_current_ood_qs_1 = (
+                self.critic_network_1(
+                    repeat_env_states[:-1],
+                    current_ood_actions[1:],
+                    current_ood_actions[1:],
+                )
+                - ood_actions_log_prob[1:]
+            )
+            next_current_ood_qs_2 = (
+                self.critic_network_2(
+                    repeat_env_states[:-1],
+                    current_ood_actions[1:],
+                    current_ood_actions[1:],
+                )
+                - ood_actions_log_prob[1:]
+            )
+
+            # Reshape
+            ood_qs_1 = tf.reshape(ood_qs_1, (T - 1, B, self.num_ood_actions, N))
+            ood_qs_2 = tf.reshape(ood_qs_2, (T - 1, B, self.num_ood_actions, N))
+            current_ood_qs_1 = tf.reshape(current_ood_qs_1, (T - 1, B, self.num_ood_actions, N))
+            current_ood_qs_2 = tf.reshape(current_ood_qs_2, (T - 1, B, self.num_ood_actions, N))
+            next_current_ood_qs_1 = tf.reshape(
+                next_current_ood_qs_1, (T - 1, B, self.num_ood_actions, N)
+            )
+            next_current_ood_qs_2 = tf.reshape(
+                next_current_ood_qs_2, (T - 1, B, self.num_ood_actions, N)
+            )
+
+            all_ood_qs_1 = tf.concat((ood_qs_1, current_ood_qs_1, next_current_ood_qs_1), axis=2)
+            all_ood_qs_2 = tf.concat((ood_qs_2, current_ood_qs_2, next_current_ood_qs_2), axis=2)
+
+            cql_loss_1 = tf.reduce_mean(
+                tf.reduce_logsumexp(all_ood_qs_1, axis=2, keepdims=False)
+            ) - tf.reduce_mean(qs_1[:-1])
+            cql_loss_2 = tf.reduce_mean(
+                tf.reduce_logsumexp(all_ood_qs_2, axis=2, keepdims=False)
+            ) - tf.reduce_mean(qs_2[:-1])
+
+            critic_loss_1 = td_error_1 + self.cql_weight * cql_loss_1
+            critic_loss_2 = td_error_2 + self.cql_weight * cql_loss_2
+            critic_loss = (critic_loss_1 + critic_loss_2) / 2
+
+            ### END CQL ###
+
+            # Policy Loss
+            policy_qs_1 = self.critic_network_1(env_states, online_actions, replay_actions)
+            policy_qs_2 = self.critic_network_2(env_states, online_actions, replay_actions)
             policy_qs = tf.minimum(policy_qs_1, policy_qs_2)
 
-            policy_loss = (
-                -(self.bc_alpha / tf.reduce_mean(tf.abs(policy_qs))) * tf.reduce_mean(policy_qs)
-                + 1e-3 * tf.reduce_mean(tf.square(online_actions))
-                + tf.reduce_mean(tf.square(online_actions - replay_actions))
-            )
+            policy_loss = -tf.reduce_mean(policy_qs) + 1e-3 * tf.reduce_mean(online_actions**2)
 
         # Train critics
         variables = (
@@ -281,6 +395,8 @@ class IDDPGBCSystem(BaseOfflineSystem):
         logs = {
             "mean_dataset_q_values": tf.reduce_mean((qs_1 + qs_2) / 2),
             "critic_loss": critic_loss,
+            "td_loss": (td_error_1 + td_error_2) / 2,
+            "cql_loss": (cql_loss_1 + cql_loss_2) / 2.0,
             "policy_loss": policy_loss,
             "mean_chosen_q_values": tf.reduce_mean((policy_qs_1 + policy_qs_2) / 2),
         }
@@ -288,7 +404,7 @@ class IDDPGBCSystem(BaseOfflineSystem):
         return logs
 
 
-@hydra.main(version_base=None, config_path="configs", config_name="iddpg_bc")
+@hydra.main(version_base=None, config_path="configs", config_name="maddpg_cql")
 def run_experiment(cfg: DictConfig) -> None:
     print(cfg)
 
@@ -314,7 +430,7 @@ def run_experiment(cfg: DictConfig) -> None:
     }
     logger = WandbLogger(project=cfg["wandb_project"], config=wandb_config)
 
-    system = IDDPGBCSystem(env, logger, **cfg["system"])
+    system = MADDPGCQLSystem(env, logger, **cfg["system"])
 
     tf.random.set_seed(cfg["seed"])
