+++ og_marl/tf2_baselines/systems/maicq.py
@@ -12,26 +12,26 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
-"""Implementation of QMIX+CQL"""
+"""Implementation of MAICQ"""
 from typing import Any, Dict, Tuple
 
-import hydra
-from omegaconf import DictConfig
-import tree
-import numpy as np
-import tensorflow as tf
-from tensorflow import Tensor
-import sonnet as snt
-from chex import Numeric
 import copy
+import hydra
+import numpy as np
+from omegaconf import DictConfig
+import sonnet as snt
+import tensorflow as tf
+import tree
+from chex import Numeric
+from tensorflow import Tensor
 
 from og_marl.environments import get_environment, BaseEnvironment
 from og_marl.loggers import BaseLogger, WandbLogger
 from og_marl.vault_utils.download_vault import download_and_unzip_vault
 from og_marl.replay_buffers import Experience, FlashbaxReplayBuffer
-from og_marl.tf2.networks import QMixer
-from og_marl.tf2.systems.base import BaseOfflineSystem
-from og_marl.tf2.utils import (
+from og_marl.tf2_baselines.networks import QMixer
+from og_marl.tf2_baselines.systems.base import BaseOfflineSystem
+from og_marl.tf2_baselines.utils import (
     batch_concat_agent_id_to_obs,
     concat_agent_id_to_obs,
     expand_batch_and_agent_dim_of_time_major_sequence,
@@ -45,9 +45,9 @@ from og_marl.tf2.utils import (
 set_growing_gpu_memory()
 
 
-class QMIXCQLSystem(BaseOfflineSystem):
+class MAICQSystem(BaseOfflineSystem):
 
-    """QMIX+CQL System"""
+    """MAICQ System"""
 
     def __init__(
         self,
@@ -61,8 +61,8 @@ class QMIXCQLSystem(BaseOfflineSystem):
         target_update_period: int = 200,
         learning_rate: float = 3e-4,
         add_agent_id_to_obs: bool = True,
-        num_ood_actions: int = 20,  # CQL
-        cql_weight: float = 3.0,  # CQL
+        icq_advantages_beta: float = 0.1,  # from MAICQ code
+        icq_target_q_taken_beta: int = 1000,  # from MAICQ code
     ):
         super().__init__(
             environment,
@@ -98,23 +98,36 @@ class QMIXCQLSystem(BaseOfflineSystem):
         self.mixer = QMixer(len(self.environment.agents), mixer_embed_dim, mixer_hyper_dim)
         self.target_mixer = QMixer(len(self.environment.agents), mixer_embed_dim, mixer_hyper_dim)
 
-        self.num_ood_actions = num_ood_actions
-        self.cql_weight = cql_weight
+        # ICQ hyper-params
+        self.icq_advantages_beta = icq_advantages_beta
+        self.icq_target_q_taken_beta = icq_target_q_taken_beta
+
+        # Policy Network
+        self.policy_network = snt.DeepRNN(
+            [
+                snt.Linear(linear_layer_dim),
+                tf.nn.relu,
+                snt.GRU(recurrent_layer_dim),
+                tf.nn.relu,
+                snt.Linear(self.environment.num_actions),
+                tf.nn.softmax,
+            ]
+        )
 
     def reset(self) -> None:
-        """Called at the start of a new episode during evaluation."""
+        """Called at the start of a new episode."""
+        # Reset the recurrent neural network
         self.rnn_states = {
-            agent: self.q_network.initial_state(1) for agent in self.environment.agents
-        }  # reset the rnn hidden states
+            agent: self.policy_network.initial_state(1) for agent in self.environment.agents
+        }
+        return
 
     def select_actions(
         self,
         observations: Dict[str, np.ndarray],
         legal_actions: Dict[str, np.ndarray],
     ) -> Dict[str, np.ndarray]:
-        observations, legal_actions = tree.map_structure(
-            tf.convert_to_tensor, (observations, legal_actions)
-        )
+        observations = tree.map_structure(tf.convert_to_tensor, observations)
         actions, next_rnn_states = self._tf_select_actions(
             observations, legal_actions, self.rnn_states
         )
@@ -132,22 +145,25 @@ class QMIXCQLSystem(BaseOfflineSystem):
     ) -> Tuple[Dict[str, Tensor], Dict[str, Tensor]]:
         actions = {}
         next_rnn_states = {}
-        for i, agent in enumerate(self.agents):
+        for i, agent in enumerate(self.environment.agents):
             agent_observation = observations[agent]
-            if self.add_agent_id_to_obs:
-                agent_observation = concat_agent_id_to_obs(agent_observation, i, len(self.agents))
+            agent_observation = concat_agent_id_to_obs(
+                agent_observation, i, len(self.environment.agents)
+            )
             agent_observation = tf.expand_dims(agent_observation, axis=0)  # add batch dimension
-            q_values, next_rnn_states[agent] = self.q_network(agent_observation, rnn_states[agent])
+            probs, next_rnn_states[agent] = self.policy_network(
+                agent_observation, rnn_states[agent]
+            )
 
             agent_legal_actions = legal_actions[agent]
-            masked_q_values = tf.where(
-                tf.cast(agent_legal_actions, "bool"),
-                q_values[0],
+            masked_probs = tf.where(
+                tf.equal(agent_legal_actions, 1),
+                probs[0],
                 -99999999,
             )
-            greedy_action = tf.argmax(masked_q_values)
 
-            actions[agent] = greedy_action
+            # Max Q-value over legal actions
+            actions[agent] = tf.argmax(masked_probs)
 
         return actions, next_rnn_states
 
@@ -157,7 +173,11 @@ class QMIXCQLSystem(BaseOfflineSystem):
         return logs  # type: ignore
 
     @tf.function(jit_compile=True)
-    def _tf_train_step(self, train_step: int, experience: Dict[str, Any]) -> Dict[str, Numeric]:
+    def _tf_train_step(
+        self,
+        train_step: int,
+        experience: Dict[str, Any],
+    ) -> Dict[str, Numeric]:
         # Unpack the batch
         observations = experience["observations"]  # (B,T,N,O)
         actions = experience["actions"]  # (B,T,N)
@@ -192,9 +212,9 @@ class QMIXCQLSystem(BaseOfflineSystem):
         target_qs_out = expand_batch_and_agent_dim_of_time_major_sequence(target_qs_out, B, N)
 
         # Make batch-major again
-        target_qs_out = switch_two_leading_dims(target_qs_out)
+        target_q_vals = switch_two_leading_dims(target_qs_out)
 
-        with tf.GradientTape() as tape:
+        with tf.GradientTape(persistent=True) as tape:
             # Unroll online network
             qs_out = unroll_rnn(self.q_network, observations, resets)
 
@@ -202,86 +222,74 @@ class QMIXCQLSystem(BaseOfflineSystem):
             qs_out = expand_batch_and_agent_dim_of_time_major_sequence(qs_out, B, N)
 
             # Make batch-major again
-            qs_out = switch_two_leading_dims(qs_out)
+            q_vals = switch_two_leading_dims(qs_out)
 
-            # Pick the Q-Values for the actions taken by each agent
-            chosen_action_qs = gather(qs_out, actions, axis=3, keepdims=False)
+            # Unroll the policy
+            probs_out = unroll_rnn(self.policy_network, observations, resets)
 
-            # Max over target Q-Values/ Double q learning
-            qs_out_selector = tf.where(
-                tf.cast(legal_actions, "bool"), qs_out, -9999999
-            )  # legal action masking
-            cur_max_actions = tf.argmax(qs_out_selector, axis=3)
-            target_max_qs = gather(target_qs_out, cur_max_actions, axis=-1)
+            # Expand batch and agent_dim
+            probs_out = expand_batch_and_agent_dim_of_time_major_sequence(probs_out, B, N)
 
-            # Q-MIXING
-            chosen_action_qs, target_max_qs, rewards = self.mixing(
-                chosen_action_qs,
-                target_max_qs,
-                env_states,
-                rewards,
-            )
+            # Make batch-major again
+            probs_out = switch_two_leading_dims(probs_out)
+
+            # Mask illegal actions
+            probs_out = probs_out * tf.cast(legal_actions, "float32")
+            probs_sum = (
+                tf.reduce_sum(probs_out, axis=-1, keepdims=True) + 1e-10
+            )  # avoid div by zero
+            probs_out = probs_out / probs_sum
+
+            action_values = gather(q_vals, actions)
+            baseline = tf.reduce_sum(probs_out * q_vals, axis=-1)
+            advantages = action_values - baseline
+            advantages = tf.nn.softmax(advantages / self.icq_advantages_beta, axis=0)
+            advantages = tf.stop_gradient(advantages)
+
+            pi_taken = gather(probs_out, actions, keepdims=False)
+            log_pi_taken = tf.math.log(pi_taken)
+
+            coe = self.mixer.k(env_states)
+
+            coma_loss = -tf.reduce_mean(coe * (len(advantages) * advantages * log_pi_taken))
+
+            # Critic learning
+            q_taken = gather(q_vals, actions, axis=-1)
+            target_q_taken = gather(target_q_vals, actions, axis=-1)
+
+            # Mixing critics
+            q_taken = self.mixer(q_taken, env_states)
+            target_q_taken = self.target_mixer(target_q_taken, env_states)
+
+            advantage_Q = tf.nn.softmax(target_q_taken / self.icq_target_q_taken_beta, axis=0)
+            target_q_taken = len(advantage_Q) * advantage_Q * target_q_taken
 
             # Compute targets
             targets = (
-                rewards[:, :-1] + (1 - terminals[:, :-1]) * self.discount * target_max_qs[:, 1:]
+                rewards[:, :-1] + (1 - terminals[:, :-1]) * self.discount * target_q_taken[:, 1:]
             )
             targets = tf.stop_gradient(targets)
 
-            # TD-Error Loss
-            td_loss = 0.5 * tf.reduce_mean(tf.square(targets - chosen_action_qs[:, :-1]))
+            # TD error
+            td_error = targets - q_taken[:, :-1]
+            q_loss = 0.5 * tf.square(td_error)
 
-            #############
-            #### CQL ####
-            #############
+            # Masking
+            q_loss = tf.reduce_mean(q_loss)
 
-            # Sample legal random actions
-            repeated_legals = tf.stack([legal_actions] * self.num_ood_actions, axis=0)
-            repeated_legals = tf.reshape(repeated_legals, (-1, A))
-            random_ood_actions = tf.random.categorical(
-                repeated_legals / tf.reduce_sum(repeated_legals, axis=-1, keepdims=True),
-                1,
-                dtype="int32",
-            )
-            random_ood_actions = tf.reshape(random_ood_actions, (self.num_ood_actions, B, T, N))
+            # Add losses together
+            loss = q_loss + coma_loss
 
-            all_mixed_ood_qs = []
-            for i in range(self.num_ood_actions):
-                # Gather
-                one_hot_indices = tf.one_hot(random_ood_actions[i], depth=qs_out.shape[-1])
-                ood_qs = tf.reduce_sum(
-                    qs_out * one_hot_indices, axis=-1, keepdims=False
-                )  # [B, T, N]
-
-                # Mixing
-                mixed_ood_qs = self.mixer(ood_qs, env_states)  # [B, T, 1]
-                all_mixed_ood_qs.append(mixed_ood_qs)  # [B, T, Ra]
-
-            all_mixed_ood_qs.append(chosen_action_qs)  # [B, T, Ra + 1]
-            all_mixed_ood_qs = tf.concat(all_mixed_ood_qs, axis=-1)
-
-            cql_loss = tf.reduce_mean(
-                tf.reduce_logsumexp(all_mixed_ood_qs, axis=-1, keepdims=True)
-            ) - tf.reduce_mean(chosen_action_qs)
-
-            #############
-            #### end ####
-            #############
-
-            # Add cql_loss to loss
-            loss = td_loss + self.cql_weight * cql_loss
-
-        # Get trainable variables
+        # Apply gradients to policy
         variables = (
+            *self.policy_network.trainable_variables,
             *self.q_network.trainable_variables,
             *self.mixer.trainable_variables,
-        )
+        )  # Get trainable variables
 
-        # Compute gradients.
-        gradients = tape.gradient(loss, variables)
+        gradients = tape.gradient(loss, variables)  # Compute gradients.
 
-        # Apply gradients.
-        self.optimizer.apply(gradients, variables)
+        self.optimizer.apply(gradients, variables)  # One optimizer for whole system
 
         # Online variables
         online_variables = (
@@ -301,33 +309,12 @@ class QMIXCQLSystem(BaseOfflineSystem):
                 dest.assign(src)
 
         return {
-            "loss": loss,
-            "cql_loss": cql_loss,
-            "td_loss": td_loss,
-            "mean_q_values": tf.reduce_mean(qs_out),
-            "mean_chosen_q_values": tf.reduce_mean(chosen_action_qs),
+            "critic_oss": q_loss,
+            "policy_loss": coma_loss,
         }
 
-    def mixing(
-        self,
-        chosen_action_qs: Tensor,
-        target_max_qs: Tensor,
-        states: Tensor,
-        rewards: Tensor,
-    ) -> Tuple[Tensor, Tensor, Tensor]:
-        """QMIX"""
-        # VDN
-        # chosen_action_qs = tf.reduce_sum(chosen_action_qs, axis=2, keepdims=True)
-        # target_max_qs = tf.reduce_sum(target_max_qs, axis=2, keepdims=True)
-        # VDN
 
-        chosen_action_qs = self.mixer(chosen_action_qs, states)
-        target_max_qs = self.target_mixer(target_max_qs, states)
-        rewards = tf.reduce_mean(rewards, axis=2, keepdims=True)
-        return chosen_action_qs, target_max_qs, rewards
-
-
-@hydra.main(version_base=None, config_path="configs", config_name="qmix_cql")
+@hydra.main(version_base=None, config_path="configs", config_name="maicq")
 def run_experiment(cfg: DictConfig) -> None:
     print(cfg)
 
@@ -353,7 +340,7 @@ def run_experiment(cfg: DictConfig) -> None:
     }
     logger = WandbLogger(project=cfg["wandb_project"], config=wandb_config)
 
-    system = QMIXCQLSystem(env, logger, **cfg["system"])
+    system = MAICQSystem(env, logger, **cfg["system"])
 
     tf.random.set_seed(cfg["seed"])
