+++ og_marl/tf2_baselines/systems/omar.py
@@ -12,26 +12,27 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
-"""Implementation of MADDPG+CQL"""
+"""Implementation of OMAR"""
 from typing import Any, Dict, Optional, Tuple
 import copy
 
 import hydra
+import tree
 import numpy as np
-import sonnet as snt
 from omegaconf import DictConfig
 import tensorflow as tf
+import sonnet as snt
 from tensorflow import Tensor
-import tree
+import tensorflow_probability as tfp
 from chex import Numeric
 
 from og_marl.environments import get_environment, BaseEnvironment
-from og_marl.tf2.networks import StateAndJointActionCritic
-from og_marl.tf2.systems.base import BaseOfflineSystem
 from og_marl.loggers import BaseLogger, WandbLogger
 from og_marl.vault_utils.download_vault import download_and_unzip_vault
 from og_marl.replay_buffers import Experience, FlashbaxReplayBuffer
-from og_marl.tf2.utils import (
+from og_marl.tf2_baselines.networks import StateAndActionCritic
+from og_marl.tf2_baselines.systems.base import BaseOfflineSystem
+from og_marl.tf2_baselines.utils import (
     batch_concat_agent_id_to_obs,
     concat_agent_id_to_obs,
     expand_batch_and_agent_dim_of_time_major_sequence,
@@ -44,9 +45,7 @@ from og_marl.tf2.utils import (
 set_growing_gpu_memory()
 
 
-class MADDPGCQLSystem(BaseOfflineSystem):
-    """Multi-Agent Deep Deterministic Policy Gradients with CQL."""
-
+class OMARSystem(BaseOfflineSystem):
     def __init__(
         self,
         environment: BaseEnvironment,
@@ -55,14 +54,22 @@ class MADDPGCQLSystem(BaseOfflineSystem):
         recurrent_layer_dim: int = 64,
         discount: float = 0.99,
         target_update_rate: float = 0.005,
-        critic_learning_rate: float = 1e-3,
+        critic_learning_rate: float = 3e-4,
         policy_learning_rate: float = 3e-4,
         add_agent_id_to_obs: bool = True,
         num_ood_actions: int = 10,  # CQL
         cql_weight: float = 3.0,  # CQL
-        cql_sigma: float = 0.2,  # CQL
+        cql_sigma: float = 0.3,  # CQL
+        omar_iters: int = 3,  # OMAR
+        omar_num_samples: int = 10,  # OMAR
+        omar_num_elites: int = 10,  # OMAR
+        omar_sigma: float = 2.0,  # OMAR
+        omar_coe: float = 0.7,  # OMAR
     ):
-        super().__init__(environment=environment, logger=logger)
+        super().__init__(
+            environment=environment,
+            logger=logger,
+        )
 
         self.add_agent_id_to_obs = add_agent_id_to_obs
         self.discount = discount
@@ -83,8 +90,8 @@ class MADDPGCQLSystem(BaseOfflineSystem):
         self.target_policy_network = copy.deepcopy(self.policy_network)
 
         # Critic network
-        self.critic_network_1 = StateAndJointActionCritic(
-            len(self.environment.agents), self.environment.num_actions
+        self.critic_network_1 = StateAndActionCritic(
+            len(self.environment.agents), self.environment.num_actions, add_agent_id_to_obs
         )  # shared network for all agents
         self.critic_network_2 = copy.deepcopy(self.critic_network_1)
 
@@ -106,6 +113,14 @@ class MADDPGCQLSystem(BaseOfflineSystem):
         self.num_ood_actions = num_ood_actions
         self.cql_weight = cql_weight
         self.cql_sigma = cql_sigma
+
+        # OMAR
+        self.omar_iters = omar_iters
+        self.omar_num_samples = omar_num_samples
+        self.omar_num_elites = omar_num_elites
+        self.omar_coe = omar_coe
+
+        self.init_omar_mu, self.init_omar_sigma = 0.0, omar_sigma
 
     def reset(self) -> None:
         """Called at the start of a new episode."""
@@ -158,9 +173,7 @@ class MADDPGCQLSystem(BaseOfflineSystem):
     def _tf_train_step(self, experience: Dict[str, Any]) -> Dict[str, Numeric]:
         # Unpack the batch
         observations = experience["observations"]  # (B,T,N,O)
-        actions = tf.clip_by_value(
-            experience["actions"], -1.0, 1.0
-        )  # (B,T,N,A) clip for omiga datasets
+        actions = experience["actions"]  # (B,T,N,A)
         env_states = experience["infos"]["state"]  # (B,T,S)
         rewards = experience["rewards"]  # (B,T,N)
         truncations = tf.cast(experience["truncations"], "float32")  # (B,T,N)
@@ -170,7 +183,7 @@ class MADDPGCQLSystem(BaseOfflineSystem):
         resets = tf.maximum(terminals, truncations)  # equivalent to logical 'or'
 
         # Get dims
-        B, T, N = actions.shape[:3]
+        B, T, N, A = actions.shape[:4]
 
         # Maybe add agent ids to observation
         if self.add_agent_id_to_obs:
@@ -178,11 +191,11 @@ class MADDPGCQLSystem(BaseOfflineSystem):
 
         # Make time-major
         observations = switch_two_leading_dims(observations)
+        resets = switch_two_leading_dims(resets)
         replay_actions = switch_two_leading_dims(actions)
         rewards = switch_two_leading_dims(rewards)
         terminals = switch_two_leading_dims(terminals)
         env_states = switch_two_leading_dims(env_states)
-        resets = switch_two_leading_dims(resets)
 
         # Unroll target policy
         target_actions = unroll_rnn(
@@ -193,8 +206,8 @@ class MADDPGCQLSystem(BaseOfflineSystem):
         target_actions = expand_batch_and_agent_dim_of_time_major_sequence(target_actions, B, N)
 
         # Target critics
-        target_qs_1 = self.target_critic_network_1(env_states, target_actions, target_actions)
-        target_qs_2 = self.target_critic_network_2(env_states, target_actions, target_actions)
+        target_qs_1 = self.target_critic_network_1(env_states, target_actions)
+        target_qs_2 = self.target_critic_network_2(env_states, target_actions)
 
         # Take minimum between two target critics
         target_qs = tf.minimum(target_qs_1, target_qs_2)
@@ -208,17 +221,17 @@ class MADDPGCQLSystem(BaseOfflineSystem):
         with tf.GradientTape(persistent=True) as tape:
             # Online critics
             qs_1 = tf.squeeze(
-                self.critic_network_1(env_states, replay_actions, replay_actions),
+                self.critic_network_1(env_states, replay_actions),
                 axis=-1,
             )
             qs_2 = tf.squeeze(
-                self.critic_network_2(env_states, replay_actions, replay_actions),
+                self.critic_network_2(env_states, replay_actions),
                 axis=-1,
             )
 
             # Squared TD-error
-            td_error_1 = tf.reduce_mean(0.5 * (targets - qs_1[:-1]) ** 2)
-            td_error_2 = tf.reduce_mean(0.5 * (targets - qs_2[:-1]) ** 2)
+            critic_loss_1 = tf.reduce_mean(0.5 * (targets - qs_1[:-1]) ** 2)
+            critic_loss_2 = tf.reduce_mean(0.5 * (targets - qs_2[:-1]) ** 2)
 
             ###########
             ### CQL ###
@@ -261,15 +274,11 @@ class MADDPGCQLSystem(BaseOfflineSystem):
             random_ood_action_log_pi = tf.math.log(0.5 ** (random_ood_actions.shape[-1]))
 
             ood_qs_1 = (
-                self.critic_network_1(repeat_env_states, random_ood_actions, random_ood_actions)[
-                    :-1
-                ]
+                self.critic_network_1(repeat_env_states, random_ood_actions)[:-1]
                 - random_ood_action_log_pi
             )
             ood_qs_2 = (
-                self.critic_network_2(repeat_env_states, random_ood_actions, random_ood_actions)[
-                    :-1
-                ]
+                self.critic_network_2(repeat_env_states, random_ood_actions)[:-1]
                 - random_ood_action_log_pi
             )
 
@@ -295,14 +304,12 @@ class MADDPGCQLSystem(BaseOfflineSystem):
                 self.critic_network_1(
                     repeat_env_states[:-1],
                     current_ood_actions[:-1],
-                    current_ood_actions[:-1],
                 )
                 - ood_actions_log_prob[:-1]
             )
             current_ood_qs_2 = (
                 self.critic_network_2(
                     repeat_env_states[:-1],
-                    current_ood_actions[:-1],
                     current_ood_actions[:-1],
                 )
                 - ood_actions_log_prob[:-1]
@@ -312,14 +319,12 @@ class MADDPGCQLSystem(BaseOfflineSystem):
                 self.critic_network_1(
                     repeat_env_states[:-1],
                     current_ood_actions[1:],
-                    current_ood_actions[1:],
                 )
                 - ood_actions_log_prob[1:]
             )
             next_current_ood_qs_2 = (
                 self.critic_network_2(
                     repeat_env_states[:-1],
-                    current_ood_actions[1:],
                     current_ood_actions[1:],
                 )
                 - ood_actions_log_prob[1:]
@@ -347,18 +352,95 @@ class MADDPGCQLSystem(BaseOfflineSystem):
                 tf.reduce_logsumexp(all_ood_qs_2, axis=2, keepdims=False)
             ) - tf.reduce_mean(qs_2[:-1])
 
-            critic_loss_1 = td_error_1 + self.cql_weight * cql_loss_1
-            critic_loss_2 = td_error_2 + self.cql_weight * cql_loss_2
-            critic_loss = (critic_loss_1 + critic_loss_2) / 2
+            critic_loss_1 += self.cql_weight * cql_loss_1
+            critic_loss_2 += self.cql_weight * cql_loss_2
 
             ### END CQL ###
 
-            # Policy Loss
-            policy_qs_1 = self.critic_network_1(env_states, online_actions, replay_actions)
-            policy_qs_2 = self.critic_network_2(env_states, online_actions, replay_actions)
-            policy_qs = tf.minimum(policy_qs_1, policy_qs_2)
+            critic_loss = (critic_loss_1 + critic_loss_2) / 2
 
-            policy_loss = -tf.reduce_mean(policy_qs) + 1e-3 * tf.reduce_mean(online_actions**2)
+            ####################
+            # OMAR Policy Loss #
+            ####################
+
+            curr_pol_out = online_actions
+
+            pred_qvals = tf.minimum(
+                self.critic_network_1(env_states, curr_pol_out),
+                self.critic_network_2(env_states, curr_pol_out),
+            )
+
+            omar_mu = tf.zeros_like(curr_pol_out) + self.init_omar_mu
+            omar_sigma = tf.zeros_like(curr_pol_out) + self.init_omar_sigma
+
+            # Repeat all tensors num_ood_actions times andadd  next to batch dim
+            observations = tf.stack(
+                [observations] * self.omar_num_samples, axis=2
+            )  # next to batch dim
+            env_states = tf.stack([env_states] * self.omar_num_samples, axis=2)  # next to batch dim
+
+            # Flatten into batch dim
+            observations = tf.reshape(observations, (T, -1, *observations.shape[3:]))
+            env_states = tf.reshape(env_states, (T, -1, *env_states.shape[3:]))
+
+            last_top_k_qvals, last_elite_acs = None, None
+            for iter_idx in range(self.omar_iters):
+                dist = tfp.distributions.Normal(omar_mu, omar_sigma)
+                cem_sampled_acs = dist.sample((self.omar_num_samples,))
+                cem_sampled_acs = tf.transpose(cem_sampled_acs, (1, 2, 0, 3, 4))
+                cem_sampled_acs = tf.clip_by_value(cem_sampled_acs, -1.0, 1.0)
+
+                formatted_cem_sampled_acs = tf.reshape(
+                    cem_sampled_acs, (T, -1, *cem_sampled_acs.shape[3:])
+                )
+
+                all_pred_qvals = tf.minimum(
+                    self.critic_network_1(env_states, formatted_cem_sampled_acs),
+                    self.critic_network_2(env_states, formatted_cem_sampled_acs),
+                )
+                all_pred_qvals = tf.reshape(all_pred_qvals, (T, B, self.omar_num_samples, N))
+                all_pred_qvals = tf.transpose(all_pred_qvals, (0, 1, 3, 2))
+                cem_sampled_acs = tf.transpose(cem_sampled_acs, (0, 1, 3, 4, 2))
+
+                if iter_idx > 0:
+                    all_pred_qvals = tf.concat((all_pred_qvals, last_top_k_qvals), axis=-1)  # type: ignore
+                    cem_sampled_acs = tf.concat((cem_sampled_acs, last_elite_acs), axis=-1)  # type: ignore
+
+                top_k_qvals, top_k_inds = tf.math.top_k(all_pred_qvals, self.omar_num_elites)
+                elite_ac_inds = tf.stack([top_k_inds] * A, axis=-2)
+                elite_acs = tf.gather(cem_sampled_acs, elite_ac_inds, batch_dims=-1)
+
+                last_top_k_qvals, last_elite_acs = top_k_qvals, elite_acs
+
+                updated_mu = tf.reduce_mean(elite_acs, axis=-1)
+                updated_sigma = tf.math.reduce_std(elite_acs, axis=-1)
+
+                omar_mu = updated_mu
+                omar_sigma = updated_sigma
+
+            top_qvals, top_inds = tf.math.top_k(all_pred_qvals, self.omar_num_elites)
+            top_ac_inds = tf.stack([top_k_inds] * A, axis=-2)
+            top_acs = tf.gather(cem_sampled_acs, top_ac_inds, batch_dims=-1)
+
+            cem_qvals = top_qvals
+            pol_qvals = pred_qvals
+            cem_acs = top_acs
+            pol_acs = tf.expand_dims(curr_pol_out, axis=-1)
+
+            candidate_qvals = tf.concat([pol_qvals, cem_qvals], -1)
+            candidate_acs = tf.concat([pol_acs, cem_acs], -1)
+
+            max_inds = tf.argmax(candidate_qvals, axis=-1)
+            max_ac_inds = tf.expand_dims(tf.stack([max_inds] * A, axis=-1), axis=-1)
+
+            max_acs = tf.gather(candidate_acs, max_ac_inds, batch_dims=-1)
+            max_acs = tf.stop_gradient(tf.reshape(max_acs, max_acs.shape[:-1]))
+
+            policy_loss = (
+                self.omar_coe * tf.reduce_mean((curr_pol_out - max_acs) ** 2)
+                - (1 - self.omar_coe) * tf.reduce_mean(pred_qvals)
+                + 1e-3 * tf.reduce_mean(curr_pol_out**2)
+            )
 
         # Train critics
         variables = (
@@ -366,11 +448,13 @@ class MADDPGCQLSystem(BaseOfflineSystem):
             *self.critic_network_2.trainable_variables,
         )
         gradients = tape.gradient(critic_loss, variables)
+        gradients, _ = tf.clip_by_global_norm(gradients, 10.0)
         self.critic_optimizer.apply(gradients, variables)
 
         # Train policy
         variables = (*self.policy_network.trainable_variables,)
         gradients = tape.gradient(policy_loss, variables)
+        gradients, _ = tf.clip_by_global_norm(gradients, 10.0)
         self.policy_optimizer.apply(gradients, variables)
 
         # Update target networks
@@ -395,16 +479,14 @@ class MADDPGCQLSystem(BaseOfflineSystem):
         logs = {
             "mean_dataset_q_values": tf.reduce_mean((qs_1 + qs_2) / 2),
             "critic_loss": critic_loss,
-            "td_loss": (td_error_1 + td_error_2) / 2,
-            "cql_loss": (cql_loss_1 + cql_loss_2) / 2.0,
+            "cql_loss": tf.reduce_mean(cql_loss_1 + cql_loss_2) / 2.0,
             "policy_loss": policy_loss,
-            "mean_chosen_q_values": tf.reduce_mean((policy_qs_1 + policy_qs_2) / 2),
         }
 
         return logs
 
 
-@hydra.main(version_base=None, config_path="configs", config_name="maddpg_cql")
+@hydra.main(version_base=None, config_path="configs", config_name="omar")
 def run_experiment(cfg: DictConfig) -> None:
     print(cfg)
 
@@ -430,7 +512,7 @@ def run_experiment(cfg: DictConfig) -> None:
     }
     logger = WandbLogger(project=cfg["wandb_project"], config=wandb_config)
 
-    system = MADDPGCQLSystem(env, logger, **cfg["system"])
+    system = OMARSystem(env, logger, **cfg["system"])
 
     tf.random.set_seed(cfg["seed"])
