+++ pyproject.toml
@@ -1,92 +1,36 @@
-[tool.ruff]
-line-length = 100
-exclude = [
-    "custom_environments"
+[build-system]
+requires = ["setuptools>=61.0"]
+build-backend = "setuptools.build_meta"
+
+[tool.setuptools.packages.find]
+include = ['og_marl*']
+
+[tool.setuptools.dynamic]
+version = {attr = "og_marl.__version__"}
+dependencies = {file = ["requirements.txt"]}
+
+[project]
+name = "og-marl"
+readme = "README.md"
+description = "Off-the-Grid MARL: Datasets and Baselines for offline MARL Research."
+authors = [
+    {name = "Claude Formanek"},
+  ]
+dynamic = ["version", "dependencies", "optional-dependencies"]
+license = {file = "LICENSE"}
+requires-python = ">=3.10"
+classifiers = [
+    "Development Status :: 3 - Alpha",
+    "Environment :: Console",
+    "Intended Audience :: Science/Research",
+    "Intended Audience :: Developers",
+    "Operating System :: OS Independent",
+    "Programming Language :: Python :: 3.10",
+    "Topic :: Scientific/Engineering :: Artificial Intelligence",
+    "Topic :: Software Development :: Libraries :: Python Modules",
+    "License :: OSI Approved :: Apache Software License",
 ]
 
-[tool.ruff.lint]
-select = ["A","B","C","D","E","F","G","N","T","W"]  # Specify list of error codes to report.
-exclude = [
-    ".tox",
-    ".git",
-    "__pycache__",
-    "build",
-    "dist",
-    "proto/*",
-    "*.pyc",
-    "*.egg-info",
-    ".cache",
-    ".eggs",
-    "custom_environments"
-]
-ignore = [
-    "A002",  # Argument shadowing a Python builtin.
-    "A003",  # Class attribute shadowing a Python builtin.
-    "D107",  # Do not require docstrings for __init__.
-    "E266",  # Do not require block comments to only have a single leading #.
-    "E731",  # Do not assign a lambda expression, use a def.
-    "B017",  # assertRaises(Exception): or pytest.raises(Exception) should be considered evil.
-    "E203",  # black and flake8 disagree on whitespace before ':'.
-
-    "D211",  # No blank lines allowed before class docstring.
-    "D213",  # Multi-line docstring summary should start at the second line.
-
-    # TEMP
-    "D100",  # Missing docstring in public module.
-    "D101",  # Missing docstring in public class.
-    "D102",  # Missing docstring in public method.
-    "D103",  # Missing docstring in public function.
-    "D104",  # Missing docstring in public package.
-    "D105",  # Missing docstring in magic method.
-    "D400",  # First line should end with a period.
-    "D401",  # First line should be in imperative mood.
-    "D404",  # First word of the first line should not be "This".
-    "D417",  # Missing argument descriptions in the docstring.
-    "D415",  # First line should end with a period, question mark, or exclamation point.
-    "T201",  # Print found?
-    "N803",  # Argument name should be lowercase.
-    "N806",  # Variable in function should be lowercase.
-
-]
-mccabe.max-complexity = 11
-pydocstyle.convention = "google"
-
-[tool.ruff.format]
-quote-style = "double"
-
-[tool.mypy]
-python_version = 3.9
-namespace_packages = true
-incremental = false
-cache_dir = ""
-warn_redundant_casts = true
-warn_return_any = true
-warn_unused_configs = true
-warn_unused_ignores = false
-allow_redefinition = true
-disallow_untyped_calls = false
-disallow_untyped_defs = true
-disallow_incomplete_defs = true
-check_untyped_defs = true
-disallow_untyped_decorators = false
-strict_optional = true
-strict_equality = true
-explicit_package_bases = true
-follow_imports = "skip"
-ignore_missing_imports = true
-exclude = "custom_environments"
-
-[[tool.mypy.overrides]]
-module = [
-    "gymnasium_robotics.*",
-    "cv2.*",
-    "tree.*",
-    "dm_env.*",
-    "supersuit.*",
-    "pettingzoo.*",
-    "var_voltage_control.*",
-    "smac.*",
-    "smacv2.*",
-    "flatland.*",
-    "flashbax.*",
-]
+[project.urls]
+"Homepage" = "https://github.com/instadeepai/og-marl"
+"Bug Tracker" = "https://github.com/instadeepai/og-marl/issues"
