+++ requirements.txt
@@ -1,12 +1,13 @@
-numpy
-dm_tree
+hydra-core==1.3.2
+numpy==1.26.4
+jax==0.4.30
+jaxlib==0.4.30
+flashbax>=0.1.2
 tensorflow[and-cuda]==2.15.1.*
-tensorflow_io
 tensorflow-probability[tf]==0.23.0
+tensorflow_io
 dm_sonnet
-wandb
 gymnasium
 requests
-jax
-hydra-core
-flashbax
+wandb
+dm_tree
